// artifacts/unb-website/src/types.ts
export type TabType = 
  | 'home' 
  | 'tentang' 
  | 'fakultas' 
  | 'pendaftaran' 
  | 'blog' 
  | 'layanan'
  // Sub-menu Tentang
  | 'profil' | 'lppm' | 'penelitian' | 'pengabdian' | 'inovasi' | 'haki' | 'books' | 'seminar'
  | 'lpm' | 'sop' | 'jobs' | 'performance' | 'tracer' | 'spmi'
  | 'ejournal' | 'bkk' | 'cooperation' | 'student_affairs' | 'pusat_studi'
  // Sub-menu Fakultas
  | 'f_agri' | 'f_econ' | 'f_forest' | 'f_science' | 'grad'
  // Halaman Program Studi
  | 'prodi_agroteknologi'
  | 'prodi_agribisnis'
  | 'prodi_manajemen'
  | 'prodi_akuntansi'
  | 'prodi_kehutanan'
  | 'prodi_biologi'
  | 'prodi_kimia'
  | 'prodi_data_sains'
  | 'prodi_magister_manajemen'
  | 'prodi_magister_agribisnis'
  | 'prodi_magister_ekonomi_pembangunan'
  // Sub-menu Layanan
  | 'simak' | 'arteri' | 'scholarship' | 'career' | 'hr' | 'repository';